chore: Move backup button to settings
