+++ mobile/lib/features/wallet/wallet_screen.dart
@@ -4,33 +4,18 @@ import 'package:get_10101/common/channel_status_notifier.dart';
 import 'package:get_10101/features/wallet/balance.dart';
 import 'package:get_10101/features/wallet/receive_screen.dart';
 import 'package:get_10101/features/wallet/onboarding/onboarding_screen.dart';
-import 'package:get_10101/features/wallet/seed_screen.dart';
 import 'package:get_10101/features/wallet/send/send_screen.dart';
 import 'package:get_10101/features/wallet/wallet_change_notifier.dart';
 import 'package:get_10101/features/wallet/wallet_theme.dart';
-import 'package:get_10101/util/preferences.dart';
 import 'package:get_10101/util/send_receive_icons.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
 
-class WalletScreen extends StatefulWidget {
+class WalletScreen extends StatelessWidget {
   static const route = "/wallet";
   static const label = "Wallet";
 
   const WalletScreen({Key? key}) : super(key: key);
-
-  @override
-  State<WalletScreen> createState() => _WalletScreenState();
-}
-
-class _WalletScreenState extends State<WalletScreen> {
-  Future<bool>? isUserSeedBackupConfirmed;
-
-  @override
-  void initState() {
-    super.initState();
-    isUserSeedBackupConfirmed = Preferences.instance.isUserSeedBackupConfirmed();
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -64,37 +49,6 @@ class _WalletScreenState extends State<WalletScreen> {
                     child: const Text("Fund Wallet"),
                   ),
                 ),
-              FutureBuilder(
-                  future: isUserSeedBackupConfirmed,
-                  builder: (BuildContext context, AsyncSnapshot<bool> snapshot) {
-                    // TODO(holzeis): Move backup seed phrase to settings
-                    // FIXME: We ignore the value of `isUserSeedBackupConfirmed` stored in
-                    // `snapshot.data` to keep the `Backup Wallet` button visible at all times for
-                    // now. We need to rework this.
-                    if (snapshot.connectionState == ConnectionState.done) {
-                      return Container(
-                        margin: const EdgeInsets.only(left: 4, right: 4),
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.stretch,
-                          children: [
-                            const SizedBox(height: 3),
-                            ElevatedButton(
-                              onPressed: () async {
-                                final res = await context.push(SeedScreen.route);
-
-                                setState(() {
-                                  isUserSeedBackupConfirmed = Future.value(res as bool);
-                                });
-                              },
-                              child: const Text("Backup Wallet"),
-                            ),
-                          ],
-                        ),
-                      );
-                    }
-                    // return an empty box if the wallet has already been backed up or the data has not been fetched yet.
-                    return const SizedBox(height: 0);
-                  }),
               const SizedBox(
                 height: 5,
               ),
